Move tweet context code out of transform functionality
